ci: harden CI/CD with cross-arch, nightly, drift guards (geth/algorand lessons)
Adopt CI/CD practices from go-ethereum and go-algorand, and fix a 32-bit
portability bug surfaced while wiring the cross-arch guard.

Workflows
- composite action .github/actions/setup: dedupe the CGO-headers + setup-go
  block repeated across ~9 jobs; cross-platform (apt/brew). Pins the Go
  toolchain via go-version-file: go.mod (was a floating "1.24").
- concurrency cancel-in-progress on all workflows for PR/branch runs; never
  cancels main pushes or scheduled corpus runs.
- ci.yml build-386: whole-module linux/386 compile as a word-size guard
  (go-xrpl's correctness is byte-exactness with rippled; an int-width
  assumption in codec/hashing silently forks account_hash/transaction_hash).
- ci.yml lint-advisory: non-blocking advisory linters (continue-on-error).
- new nightly.yml: offer suite under -race (the PR gate skips -race there),
  native macOS/arm64 leg, linux/386 test *run*, longer consensus soak,
  informational coverage, and a benchmark smoke (must-still-run).

Quality gates
- depguard rule: public packages may not import internal/ (0 violations
  today; locks the boundary). cmd/scripts/tests exempt.
- .golangci-warnings.yml: advisory linter set for the non-blocking job.
- .codecov.yml: coverage informational only, ignores generated/test code.
- internal/tx/result_lockstep_test.go: fail if any result code in resultNames
  is missing from / renumbered against definitions.json TRANSACTION_RESULTS
  (the empty-meta-blob silent-fork class from #726).

Bug fix (found by the 386 guard)
- codec/binarycodec EncodeForSigningBatch compared len(txIDs) (int) against
  math.MaxUint32; the untyped constant overflows int on 32-bit, so the whole
  module failed to compile for any 32-bit target. Compare in uint64 (same
  intent, portable). Behaviour on 64-bit is unchanged. Plus a 32-bit
  test-constant overflow in amount_test.go.

Verified: lockstep + edited codec tests pass, depguard 0 issues, both
golangci configs verify, whole module builds for linux/386, actionlint exit 0.

+++ codec/binarycodec/types/amount_test.go
@@ -933,7 +933,7 @@ func TestAmount_FromJson(t *testing.T) {
 		},
 		{
 			name:     "fail - invalid amount type",
-			input:    10000000000000000,
+			input:    int64(10000000000000000),
 			expected: nil,
 			err:      errors.New("invalid amount type"),
 			expPass:  false,
